refactor(sirn): share HTTP test helpers via tests/common module
